изолировал логику инфы о юзере
